# api/cdn-api/swagger/docs/swagger.yaml
basePath: /
definitions:
  api.CloudflareAccountDetailRequestBody:
    properties:
      id:
        type: string
    type: object
  api.CloudflareAccountListAccountsRequestBody:
    properties:
      direction:
        type: string
      page:
        type: string
      per_page:
        type: string
    type: object
  api.CloudflareZonePurgeRequestBody:
    properties:
      path:
        type: string
    type: object
...
host: localhost:8080
info:
  contact:
    email: fancyshenmue@gmail.com
    name: Charles Hsu
  description: CDN RESTful API
  license: {}
  termsOfService: http://swagger.io/terms/
  title: CDN RESTful API
  version: "1.0"
paths:
  /api/v1/cloudflare/account/details:
    post:
      consumes:
      - application/json
      operationId: "1"
      parameters:
      - description: body
        in: body
        name: params
        required: true
        schema:
          $ref: '#/definitions/api.CloudflareAccountDetailRequestBody'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: string
      summary: Cloudflare Account Details
      tags:
      - Cloudfalre (Account)
  /api/v1/cloudflare/account/list/accounts:
    post:
      consumes:
      - application/json
      operationId: "1"
      parameters:
      - description: body
        in: body
        name: params
        required: true
        schema:
          $ref: '#/definitions/api.CloudflareAccountListAccountsRequestBody'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: string
      summary: Cloudflare Account List Accounts
      tags:
      - Cloudfalre (Account)
...
  /api/v1/cloudflare/zone/purge/all/files:
    post:
      consumes:
      - application/json
      operationId: "1"
      parameters:
      - description: body
        in: body
        name: params
        required: true
        schema:
          $ref: '#/definitions/api.CloudflareZonePurgeRequestBody'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: string
      summary: Cloudflare Zone Purge All Files
      tags:
      - Cloudfalre (Zones)
  /api/v1/cloudflare/zone/purge/files/by/url:
    post:
      consumes:
      - application/json
      operationId: "1"
      parameters:
      - description: body
        in: body
        name: params
        required: true
        schema:
          $ref: '#/definitions/api.CloudflareZonePurgeRequestBody'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: string
      summary: Cloudflare Zone Purge Files by URL
      tags:
      - Cloudfalre (Zones)
...
swagger: "2.0"